requirements.txt
langgraph>=0.2.0
langchain>=0.3.0
langchain-groq>=0.1.0
langchain-community>=0.3.0
groq>=0.9.0
tavily-python>=0.3.0
yfinance>=0.2.0
plotly>=5.18.0
reportlab>=4.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
streamlit>=1.35.0
pydantic>=2.0.0
python-dotenv>=1.0.0
pandas>=2.0.0
numpy>=1.26.0
requests>=2.31.0
